Refactor to place ports on component dynamically based on component definition
