+++ src/domain-model/OMPort.ts
@@ -1,0 +1,12 @@
+export interface OMPort {
+  id: string;
+  placement: OMPortPlacement;
+  rotation: number;
+}
+
+export interface OMPortPlacement {
+  bottomLeft: {
+    x: number;
+    y: number;
+  };
+}
